+++ ot/__init__.py
@@ -37,6 +37,7 @@ from . import solvers
 from . import gaussian
 from . import lowrank
 from . import gmm
+from . import sgot
 
 # OT functions
 from .lp import (
@@ -85,9 +86,12 @@ from .solvers import solve, solve_gromov, solve_sample
 from .lowrank import lowrank_sinkhorn
 
 from .batch import solve_batch, solve_sample_batch, solve_gromov_batch, dist_batch
+from .bsp import compute_bspot_bijection, merge_bijections
+
 
 # utils functions
 from .utils import dist, unif, tic, toc, toq
+
 
 __version__ = "0.9.7.dev0"
 
@@ -151,6 +155,7 @@ __all__ = [
     "factored",
     "lowrank",
     "gmm",
+    "sgot",
     "binary_search_circle",
     "wasserstein_circle",
     "semidiscrete_wasserstein2_unif_circle",
@@ -163,4 +168,6 @@ __all__ = [
     "solve_gromov_batch",
     "solve_sample_batch",
     "dist_batch",
+    "compute_bspot_bijection",
+    "merge_bijections",
 ]